Create Xdmf branch for ParaView
A new ParaView CVS repository was created on 2004-03-12 to re-organize
the source tree layout.  The repository used server-side symbolic
links under Utilities/Xdmf within ParaView:

  Utilities/Xdmf/vtk    -> .../Xdmf/vtk
  Utilities/Xdmf/CMake  -> .../Xdmf/CMake
  Utilities/Xdmf/libsrc -> .../Xdmf/libsrc

to share the directories with the Xdmf repository.  The repository
also made a copy of some files

  Utilities/Xdmf/Ice/*  <= .../Xdmf/Ice/libsrc/*

The approach essentially created a fork of Xdmf with only a subset of
its tree.  This commit was manufactured during conversion from CVS to
represent the fork as a branch.