Create Xdmf2 branch for ParaView
The ParaView CVS repository was modified on 2007-07-09 to add
server-side symbolic links under Utilities/Xdmf2 within ParaView:

  Utilities/Xdmf2/vtk    -> .../Xdmf/vtk
  Utilities/Xdmf2/CMake  -> .../Xdmf/CMake
  Utilities/Xdmf2/libsrc -> .../Xdmf/libsrc

and share the directories with the Xdmf repository.

The approach essentially created a fork of Xdmf with only a subset of
its tree.  This commit was manufactured during conversion from CVS to
represent the fork as a branch.