Copied over testing suite that I made on an old Xdmf2 repo.
The testing suite should make it easier to add new tests. It also
makes use of a single .cmake.in file that gets reconfigured for
each test, instead of creating a new .cmake.in file for each test.
Look into the core/CMake/Use*.cmake files to see how to add new tests.