ENH: Complete hyperslab reading.
Still an issue with invalid read of size 1 in hdf5 in valgrind with the new test.
This has to do with overwriting a dataset with a smaller sized dataset when calling
H5Dset_extent(). It is deep within hdf5 when dealing with chunked data. No errors
from hdf5 returned however.