BUG: Fix bug found by Di when dealing with text nodes in XML.
The code to deal with "Content" stored in XML Text Nodes was previously
rewritten to avoid use of xmlNodeGetContent() which recursively combines
all children node contents to the content of the current node, which was
undesired behavior. I was rewritten to find the first text node and return
the content of that node, but this was insufficient as sometimes the first
text node is whitespace. The code now finds the first text node containing
content, which works for all Xdmf and NetDMF files we have come across.