ENH: Allow arbitrary elements to be partitioned by metis.
The new version of metis uses a different format to supply connectivity
information to the partitioning algorithm. The new format allows generic
element types to be passed (as well as mixed topology formats). These changes
expand the number of element types we can partition in metis (this was tested
on high-order hexahedron, linear hexahedron, and polyline topologies)