BUG: Add typemap for std::vector<shared_ptr<XdmfItem> >
When using XPaths from Python, the reader returns a vector of XdmfItems
fitting the XPath description. However, it only returns the base class,
which is useless. This commit adds a typemap in swig to convert the base
classes to the appropriate types in a python list, allowing utilization
of XPaths to from Python.