ENH: Adjust exodus writer to utilize material blocks.
Previously, when converting from xdmf to exodus, spatial partitioning (metis)
in xdmf was written as element blocks in exodus. However, exodus uses element
blocks to specify different materials and not as spatial partitioning for
parallel execution. We now write materials to blocks in exodus, indicated by
Block node sets in xdmf. This now makes the conversion of xdmf -> exodus the
exact inverse of the exodus -> xdmf function.