ENH: Enable flexibility in specifying dimensionality of light data description.
Previously, for HDF5 datasets, the number of dimensions in the light data
description had to match the number of dimensions in the heavy data file in
order to properly select a hyperslab. To support older files and be more
flexible, we now allow the number of dimensions to be different as long as
we are reading the entire dataset.