BUG: Fix bug causing missed shared vertices in topology converter.
Checking for shared faces is not enough when inserting vertices to
generate higher order elements. Matching faces depends on the
order of traversal of the elements. Now we check for shared edges in addition
to faces - this solves the traversal ordering problem as we will pick up cases
where two elements share an edge before the faces can be matched. Before, this
caused duplicate points to be inserted on the shared edge.