backport changes from VTK 7.0.0
Some style. Some fixes, especially for windows with the exports and template changes.

More so adapt this to work well as a subdirectory in a larger project. Don't force settings
that the parent project owns, allow parent project to turn on tests without xdmf's tests and
make the name of the library an option so that parent can have two xdmfs.