fix windows and make static with shared optional
windows fix is mostly the change to XdmfCoreReader.
On kitware's dashboards machines these template defs cause
the compiler to redefine parts of stdlib. But also
the static/shared lib had bad declspecs

Besides that I added a new option XDMF_STATIC_AND_SHARED,
that enables or disables the recent feature where compilation
produces both static and shared libraries simultaneously. The default
is ON in *Nix, and disabled on windows. If/when the windows
declspec issues are resolved we can easily enable it there.