fix windows compilation
I have yet to find the case where this was necessary for Andrew.
With the __int64 block removed, it compiles on vs2013 and vs2016
64 bit, with it comp fails because of redefinition.

A simple test case below shows that the behavior is the same on
32 bit. If we find out where the test case compiles, we can add
it back in with appropriate guards to make it compile widely.

template <typename T> void setArrayType()
{ std::cout << "other" << std::endl; }
template <> void setArrayType<long long>()
{ std::cout << "long long" << std::endl; }
template <> void setArrayType<__int64>()
{ std::cout << "__int64" << std::endl; }
int main(int argc, char *argv[])
{ return 0; }